Guessing Game
This is a guessing game where users will need to guess a number in a number of guesses specified by the user. The range of numbers and the number of games played are also specified by the user. After each guess, the user will be told whether their guess is too high or too low.At they end, they will be told how many games they won. There are still several minor changes required. For example, if the user enters a lower limit higher than the upper limit, it should be flipped. Similarly, if the user enters a letter instead of a number, the program should prompt them to enter a number instead of producing an error message.